additional changes

+++ runapp.bat
@@ -1,2 +1,11 @@
 call setenv.bat
-"%JAVA_HOME%"\bin\java -cp .\bin com.tokbox.bookstore.BookStoreMainApp
+@echo off
+if not exist ".\bin\com\tokbox\bookstore\BookStoreMainApp.class" (
+  
+  echo.
+  echo Please compile application first. Run compile.bat or ant
+  
+) else (
+  echo.
+  "%JAVA_HOME%"\bin\java -cp .\bin com.tokbox.bookstore.BookStoreMainApp
+)
